@@ -1,6 +1,6 @@
 {
     "name": "kiribi_ususama",
-    "version": "9.1.0-5.1.3",
+    "version": "9.1.1-5.1.3",
     "description": "Gulp を使用した小・中規模Webサイト制作用のフレームワークです。",
     "main": "gulpfile.js",
     "scripts": {
@@ -50,29 +50,29 @@
     "author": "Arm Band",
     "license": "ISC",
     "dependencies": {
+        "@arm-band/alpha1-bootstrap-jp-sutra": "^0.0.6",
         "@fortawesome/fontawesome-free": "^6.1.1",
         "bootstrap": "^5.1.3",
-        "@arm-band/alpha1-bootstrap-jp-sutra": "^0.0.6",
         "bowser": "^2.11.0",
         "brand-colors": "^2.1.1",
-        "luminous-lightbox": "^2.4.0",
         "list.js": "^2.3.1",
-        "swiper": "^8.1.4"
+        "luminous-lightbox": "^2.4.0",
+        "swiper": "^8.3.1"
     },
     "devDependencies": {
         "axios": "^0.27.2",
-        "browser-sync": "^2.27.9",
+        "browser-sync": "^2.27.10",
         "clean-webpack-plugin": "^4.0.0",
         "cookie-parser": "^1.4.6",
         "copyfiles": "^2.4.1",
         "debug": "^4.3.4",
-        "dotenv": "^16.0.0",
-        "ejs": "^3.1.7",
-        "eslint": "^8.14.0",
+        "dotenv": "^16.0.1",
+        "ejs": "^3.1.8",
+        "eslint": "^8.19.0",
         "eslint-config-prettier": "^8.5.0",
         "express": "^4.18.1",
         "front-matter": "^4.0.2",
-        "ftp-deploy": "^2.4.2",
+        "ftp-deploy": "^2.4.3",
         "gulp": "^4.0.2",
         "gulp-autoprefixer": "^8.0.0",
         "gulp-connect-php": "^1.0.3",
@@ -80,7 +80,7 @@
         "gulp-ejs": "^5.1.0",
         "gulp-frontnote": "^2.0.1",
         "gulp-htmlmin": "^5.0.1",
-        "gulp-imagemin": "^7.1.0",
+        "gulp-imagemin": "7.1.0",
         "gulp-notify": "^4.0.0",
         "gulp-plumber": "^1.2.1",
         "gulp-rename": "^2.0.0",
@@ -90,32 +90,33 @@
         "htmllint-cli": "^0.0.7",
         "http-errors": "^2.0.0",
         "imagemin-gifsicle": "^7.0.0",
-        "imagemin-mozjpeg": "^9.0.0",
+        "imagemin-mozjpeg": "9.0.0",
         "imagemin-pngquant": "^9.0.2",
-        "imagemin-svgo": "^9.0.0",
+        "imagemin-svgo": "9.0.0",
         "lodash.sortby": "^4.7.0",
-        "marked": "^4.0.14",
+        "marked": "^4.0.18",
         "mkdirp": "^1.0.4",
         "morgan": "^1.10.0",
         "npm-run-all": "^4.1.5",
         "opener": "^1.5.2",
-        "prettier": "^2.6.2",
+        "prettier": "^2.7.1",
         "rimraf": "^3.0.2",
         "rss": "^1.2.2",
-        "sass": "^1.51.0",
-        "stylelint": "^14.8.1",
+        "sass": "^1.53.0",
+        "stylelint": "^14.9.1",
         "stylelint-config-prettier": "^9.0.3",
-        "stylelint-config-recommended-scss": "^6.0.0",
-        "stylelint-config-standard": "^25.0.0",
-        "stylelint-config-standard-scss": "3.0.0",
-        "stylelint-scss": "^4.2.0",
-        "terser-webpack-plugin": "^5.3.1",
-        "webpack": "^5.72.0",
+        "stylelint-config-recommended-scss": "^7.0.0",
+        "stylelint-config-standard": "^26.0.0",
+        "stylelint-config-standard-scss": "^4.0.0",
+        "stylelint-scss": "^4.3.0",
+        "terser-webpack-plugin": "^5.3.3",
+        "webpack": "^5.73.0",
         "webpack-stream": "^7.0.0",
-        "yaml": "^2.0.0"
+        "yaml": "^2.1.1"
     },
     "resolutions": {
-        "gifsicle": "5.3.0"
+        "gifsicle": "5.2.1",
+        "autoprefixer": "10.4.5"
     },
     "engines": {
         "node": ">=16.14.0",
